don't execute ingest API tests on build #2301
"TestIngestIT" matches "Test" so it was be executed by Maven and Travis.
